# portal.openapi.yaml
openapi: 3.1.0
info:
  title: Portal Tattoo API
  version: 1.0.0
  description: Public and administrative HTTP contract for Portal Tattoo.
servers:
  - url: /api/v1
...
security: []
paths:
...
  /public/editorial:
    get:
      tags: [Public Editorial]
      operationId: listPublishedEditorial
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/EditorialType'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Published editorial feed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EditorialPage'}
  /public/editorial/{slug}:
    get:
      tags: [Public Editorial]
      operationId: getPublishedEditorial
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Published content
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicEditorialContent'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /admin/products:
    get:
      tags: [Admin Catalog]
      security: [{oidc: []}]
      operationId: listAdminProducts
      x-required-capability: catalog.read
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/AdminProductSummary'}
    post:
      tags: [Admin Catalog]
      security: [{oidc: []}]
      operationId: createProduct
      x-required-capability: catalog.write
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateProductRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Validation'}
  /admin/products/{id}:
    get:
      tags: [Admin Catalog]
      security: [{oidc: []}]
      operationId: getAdminProduct
      x-required-capability: catalog.read
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Admin Catalog]
      security: [{oidc: []}]
      operationId: updateAdminProduct
      x-required-capability: catalog.write
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateProductRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminProduct'}
        '409': {$ref: '#/components/responses/ConcurrentModification'}
        '422': {$ref: '#/components/responses/Validation'}
  /admin/claims:
    post:
      tags: [Admin Knowledge]
      security: [{oidc: []}]
      operationId: recordClaim
      x-required-capability: claim.write
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordClaimRequest'}
      responses:
        '201':
          description: Claim recorded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Claim'}
        '422': {$ref: '#/components/responses/Validation'}
  /admin/claims/{id}:
    get:
      tags: [Admin Knowledge]
      security: [{oidc: []}]
      operationId: getAdminClaim
      x-required-capability: claim.read
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Claim detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Claim'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/canonical-proposals:
    post:
      tags: [Admin Knowledge]
      security: [{oidc: []}]
      operationId: createCanonicalProposal
      x-required-capability: canonical.propose
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCanonicalProposalRequest'}
      responses:
        '201':
          description: Proposal created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CanonicalProposal'}
  /admin/canonical-proposals/{id}/decision:
    post:
      tags: [Admin Knowledge]
      security: [{oidc: []}]
      operationId: decideCanonicalProposal
      x-required-capability: canonical.decide
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CanonicalDecisionRequest'}
      responses:
        '200':
          description: Decision applied
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CanonicalDecisionResult'}
        '409': {$ref: '#/components/responses/ConcurrentModification'}
        '422': {$ref: '#/components/responses/Validation'}
...
  /admin/editorial:
    post:
      tags: [Admin Editorial]
      security: [{oidc: []}]
      operationId: createEditorialDraft
      x-required-capability: editorial.write
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateEditorialRequest'}
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEditorialContent'}
  /admin/editorial/{id}:
    get:
      tags: [Admin Editorial]
      security: [{oidc: []}]
      operationId: getAdminEditorial
      x-required-capability: editorial.read
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Editorial content detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEditorialContent'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Admin Editorial]
      security: [{oidc: []}]
      operationId: removeEditorial
      x-required-capability: editorial.write
      description: Remove permanentemente o conteúdo editorial.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Removed
        '409': {$ref: '#/components/responses/ConcurrentModification'}
  /admin/editorial/{id}/publish:
    post:
      tags: [Admin Editorial]
      security: [{oidc: []}]
      operationId: publishEditorial
      x-required-capability: editorial.publish
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEditorialContent'}
        '409': {$ref: '#/components/responses/ConcurrentModification'}
  /admin/editorial/{id}/unpublish:
    post:
      tags: [Admin Editorial]
      security: [{oidc: []}]
      operationId: unpublishEditorial
      x-required-capability: editorial.publish
      description: Retira o conteúdo publicado do ar, retornando-o ao estado DRAFT.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Unpublished (DRAFT)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminEditorialContent'}
        '409': {$ref: '#/components/responses/ConcurrentModification'}
  /admin/listings/{id}/prices:
    post:
      tags: [Admin Commerce]
      security: [{oidc: []}]
      operationId: recordPriceObservation
      x-required-capability: commerce.manage
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordPriceRequest'}
      responses:
        '201':
          description: Price observation appended
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PriceObservation'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
    Slug:
      name: slug
      in: path
      required: true
      schema: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 24}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 200}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Aggregate version, formatted as a quoted integer ETag.
      schema: {type: string, pattern: '^"[0-9]+"$'}
  responses:
...
    NotFound:
      description: Resource not found
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Conflict:
      description: Conflict
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    ConcurrentModification:
      description: Aggregate changed since client read
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Validation:
      description: Validation failed
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ValidationProblem'}}}
  schemas:
    Uuid:
      type: string
      format: uuid
    Problem:
      type: object
      required: [type,title,status,code,correlationId]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        correlationId: {$ref: '#/components/schemas/Uuid'}
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field,code]
                properties:
                  field: {type: string}
                  code: {type: string}
                  message: {type: string}
    PageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: [string,'null']}
        hasMore: {type: boolean}
...
    CreateProductRequest:
      type: object
      required: [manufacturerId,productTypeKey,name]
      additionalProperties: false
      properties:
        manufacturerId: {$ref: '#/components/schemas/Uuid'}
        brandId: {$ref: '#/components/schemas/Uuid'}
        productTypeKey: {type: string}
        name: {type: string, minLength: 1}
        slug: {type: string}
        modelCode: {type: string}
    UpdateProductRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name: {type: string}
        brandId: {type: [string,'null'], format: uuid}
        modelCode: {type: [string,'null']}
        lifecycle: {type: string, enum: [ANNOUNCED,ACTIVE,DISCONTINUED,LEGACY,UNKNOWN]}
    AdminProductSummary:
      type: object
      required: [id,name,slug,manufacturerName,lifecycle,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        name: {type: string}
        slug: {type: string}
        manufacturerName: {type: string}
        lifecycle: {type: string}
        version: {type: integer}
    AdminProduct:
      allOf:
        - $ref: '#/components/schemas/AdminProductSummary'
        - type: object
          properties:
            claimsSummary: {type: object, additionalProperties: true}
            canonicalFacts: {type: array, items: {$ref: '#/components/schemas/CanonicalFact'}}
            media: {type: array, items: {$ref: '#/components/schemas/MediaAsset'}}
...
    RecordClaimRequest:
      type: object
      required: [subjectType,subjectId,propertyKey,value,claimantType]
      additionalProperties: false
      properties:
        subjectType: {type: string}
        subjectId: {$ref: '#/components/schemas/Uuid'}
        propertyKey: {type: string}
        value: {}
        claimantType: {type: string}
        claimantId: {type: string}
        sourceSnapshotId: {$ref: '#/components/schemas/Uuid'}
        sourceUrl: {type: string, format: uri}
        observedAt: {type: string, format: date-time}
        confidence: {type: number, minimum: 0, maximum: 1}
    Claim:
      type: object
      required: [id,subjectType,subjectId,propertyKey,value,status,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        subjectType: {type: string}
        subjectId: {$ref: '#/components/schemas/Uuid'}
        propertyKey: {type: string}
        value: {}
        status: {type: string}
        version: {type: integer}
    CanonicalFact:
      type: object
      required: [id,propertyKey,value,validFrom,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        propertyKey: {type: string}
        value: {}
        validFrom: {type: string, format: date-time}
        validTo: {type: [string,'null'], format: date-time}
        version: {type: integer}
    CreateCanonicalProposalRequest:
      type: object
      required: [subjectType,subjectId,propertyKey,proposedValue,evidenceIds]
      properties:
        subjectType: {type: string}
        subjectId: {$ref: '#/components/schemas/Uuid'}
        propertyKey: {type: string}
        proposedValue: {}
        evidenceIds:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/Uuid'}
    CanonicalProposal:
      type: object
      required: [id,status,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        status: {type: string}
        version: {type: integer}
    CanonicalDecisionRequest:
      type: object
      required: [decision,reason,expectedVersion]
      properties:
        decision: {type: string, enum: [APPROVE,REJECT]}
        reason: {type: string, minLength: 3}
        expectedVersion: {type: integer, minimum: 1}
    CanonicalDecisionResult:
      type: object
      required: [proposalId,decision]
      properties:
        proposalId: {$ref: '#/components/schemas/Uuid'}
        decision: {type: string, enum: [APPROVE,REJECT]}
        canonicalFact: {$ref: '#/components/schemas/CanonicalFact'}
    MediaRightsStatus:
      type: string
      enum: [UNKNOWN,PENDING,PERMITTED,RESTRICTED,EXPIRED,TAKEDOWN]
...
    MediaAsset:
      type: object
      required: [id,kind,mimeType,rightsStatus,status,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        kind: {type: string}
        mimeType: {type: string}
        rightsStatus: {$ref: '#/components/schemas/MediaRightsStatus'}
        status: {type: string}
        version: {type: integer}
...
    EditorialType:
      type: string
      enum: [NEWS,BLOG,EVENT,TECHNICAL_ARTICLE,NOTICE]
    CreateEditorialRequest:
      type: object
      required: [contentType,title,body]
      properties:
        contentType: {$ref: '#/components/schemas/EditorialType'}
        title: {type: string}
        subtitle: {type: string}
        summary: {type: string}
        body: {type: object, additionalProperties: true}
    AdminEditorialContent:
      type: object
      required: [id,contentType,slug,title,status,version]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        contentType: {$ref: '#/components/schemas/EditorialType'}
        slug: {type: string}
        title: {type: string}
        status: {type: string}
        version: {type: integer}
    PublicEditorialContent:
      type: object
      required: [id,contentType,slug,title,body,publishedAt]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        contentType: {$ref: '#/components/schemas/EditorialType'}
        slug: {type: string}
        title: {type: string}
        summary: {type: [string,'null']}
        body: {type: object, additionalProperties: true}
        publishedAt: {type: string, format: date-time}
    EditorialPage:
      type: object
      required: [items,meta]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/PublicEditorialContent'}}
        meta: {$ref: '#/components/schemas/PageMeta'}
    RecordPriceRequest:
      type: object
      required: [amount,currency,observedAt]
      properties:
        amount: {type: number, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3}
        availability: {type: string}
        shippingAmount: {type: number, minimum: 0}
        observedAt: {type: string, format: date-time}
        snapshotId: {$ref: '#/components/schemas/Uuid'}
    PriceObservation:
      type: object
      required: [id,amount,currency,observedAt]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        amount: {type: number}
        currency: {type: string}
        observedAt: {type: string, format: date-time}
...